Fixed up test class and test case naming
