+++ Source/MoreDotNet.Test/Extensions/Common/ObjectExtensions/ToDictionaryTests.cs
@@ -4,24 +4,8 @@
     using MoreDotNet.Extensions.Common;
     using Xunit;
 
-    public class TypeCastTests
+    public class ToDictionaryTests
     {
-        [Fact]
-        public void As_StringObjectAsString_ShouldReturnString()
-        {
-            object input = "Value";
-            var result = input.As<string>();
-            Assert.IsType<string>(result);
-        }
-
-        [Fact]
-        public void As_IntAsString_ShouldReturnNull()
-        {
-            object input = 3;
-            var result = input.As<string>();
-            Assert.Null(result);
-        }
-
         [Fact]
         public void ToDictionary_Null_ShouldThrowArgumentNullException()
         {
